+++ Changes.md
@@ -16,6 +16,7 @@
 - when using message-stream, don't send default options #3290
 - auth_base: enable disabling constrain_sender at runtime #3298
 - auth_base: skip constrain_sender when auth user has no domain #3319
+- rcpt_to.host_list: add connection ID to log messages #3322
 - connection: support IPv6 when setting remote.is_private #3295
   - in setTLS, replace forEach with for...of
   - NOTE: remove a handful of 3.0 sunset property names #3315
@@ -29,8 +30,9 @@
   - now: [A63B62DF-F3B8-4096-8996-8CE83494A188.1.1] [outbound] Failed to get socket: connect ECONNREFUSED 172.16.16.14:25
   - shorter logger syntax: logger.loginfo -> logger.info
   - outbound: remove log prefixes of `[outbound] `, no longer needed
-- logger(connection, plugins, hmail), refactor copying in log functions #3322
 - logger: don't load outbound (race condition). Instead, set name property #3322
+- logger: extend add_log_methods to Classes (connection, plugins, hmail) #3322
+- logger: when logging via `logger` methods, use short names #3322
 - outbound: delete try_deliver_host. Use net_utils to resolve MX hosts to IPs #3322
 - outbound: remove config setting ipv6_enabled #3322
 - outbound: remove undocumented use of send_email with arity of 2. #3322